Add the dumbest "chaos" causer that can work
Called the "muddle marmoset" as it is not very chaotic, and pretty
small. All it does is execute os:cmd. See the riakc_nhs.config for an
example of it stopping and starting a remote node.

If this were to be generally useful it would probably have more
commands, and a way to change config/restart workers so they didn't crash.